リポジトリの一覧を取得するAPIを整理
- URLを修正
- 不要なOpenAIへのアクセスを削除
